% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-export.R
\name{print.generalKSStat}
\alias{print.generalKSStat}
\alias{getStatName}
\alias{getStatValue}
\alias{getSampleSize}
\alias{getPvalue}
\title{Class methods for the generalKSStat S3 object}
\usage{
\method{print}{generalKSStat}(x, ...)

getStatName(x)

getStatValue(x)

getSampleSize(x)

getPvalue(x)
}
\arguments{
\item{x}{the generalKSStat S3 object}

\item{...}{Ignored.}
}
\value{
print: invisible \code{x}
other fucntions: a numeric value
}
\description{
Class methods for the generalKSStat S3 object
}
\examples{

## Generate samples
x <- rbeta(10, 1, 2)

## Perform KS test
GKSStat(x = x, statName = "KS")

}
