% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-export.R
\name{print.generalKSStat}
\alias{print.generalKSStat}
\title{print function for the generalKSStat S3 object}
\usage{
\method{print}{generalKSStat}(x, ...)
}
\arguments{
\item{x}{the generalKSStat S3 object}
}
\value{
invisible x
}
\description{
print function for the generalKSStat S3 object
}
\examples{

## Generate samples
x <- rbeta(10, 1, 2)

## Perform KS test
GKSStat(x = x, statName = "KS")

}
